Tectrics/requirements.txt
asgiref==3.7.2
contourpy==1.2.1
cycler==0.12.1
Django==4.2.11
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
networkx==3.3
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
